backend/api/requirements.txt
asyncpg~=0.29.0
fastapi~=0.105.0
pydantic~=2.5.2
python-dotenv~=1.0.0
uvicorn[standard]~=0.24.0.post1
motor~=3.3.2
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
python-multipart~=0.0.9
bcrypt~=4.0.1